CHanging to Bigdecimal the currency

+++ src/main/java/com/playtomic/tests/wallet/entity/Transaction.java
@@ -6,6 +6,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -13,16 +14,16 @@ import java.util.Date;
 @NoArgsConstructor
 @AllArgsConstructor
 public class Transaction {
-    private double amount;
+    private BigDecimal amount;
     private String status;
     private String concept;
-    private double relativeBalance;
-    private double currentBalance;
+    private BigDecimal relativeBalance;
+    private BigDecimal currentBalance;
     private String paymentId;
     private String cardAlias;
     private Date transactionDate;
 
-    public Transaction(double amount, String concept, TransactionStatus status) {
+    public Transaction(BigDecimal amount, String concept, TransactionStatus status) {
         this.amount = amount;
         this.concept = concept;
         this.status = status.name();
