+++ src/main/java/com/playtomic/tests/wallet/entity/Wallet.java
@@ -6,6 +6,7 @@ import lombok.NoArgsConstructor;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -18,13 +19,13 @@ public class Wallet {
     @Id
     private String id;
     private String playerName;
-    private double balance;
+    private BigDecimal balance;
 
     private List<Card> cards;
 
     private List<Transaction> transactions;
 
-    public Wallet(String playerName, double balance) {
+    public Wallet(String playerName, BigDecimal balance) {
         this.playerName = playerName;
         this.balance = balance;
         this.cards = new ArrayList<>();
