+++ src/main/java/com/playtomic/tests/wallet/model/PaymentOrderDto.java
@@ -3,6 +3,8 @@ package com.playtomic.tests.wallet.model;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class PaymentOrderDto {
     @JsonProperty("walletId")
@@ -10,9 +12,9 @@ public class PaymentOrderDto {
     @JsonProperty("cardAlias")
     private String cardAlias;
     @JsonProperty("amount")
-    private double amount;
+    private BigDecimal amount;
 
-    public PaymentOrderDto(String walletId, String cardAlias, double amount) {
+    public PaymentOrderDto(String walletId, String cardAlias, BigDecimal amount) {
         this.walletId = walletId;
         this.cardAlias = cardAlias;
         this.amount = amount;
